chapter_14: added exercise 1, 2 and 3
The first exercise could not be made exactely right, because it was programmed on a mac. Mac's do not have a enviromental variable for the full name of the logged in person, only the username. E.g. for me it is "nickespig" not "Nick Espig". On windows it is possilbe because there is a variable named "USERNAME" which returns the whole name.